Refactor after monkeypatch
