+++ tests/test_after.py
@@ -306,7 +306,7 @@ class TestAfter:
             'envlist = py36\n'
         )
         ini = py.iniconfig.IniConfig('', data=inistr)
-        assert not after_config_matches(['py36'], ini)
+        assert not after_config_matches(ini, ['py36'])
 
     def test_after_config_matches_toxenv_match(self, capsys):
         """Test that it works using the legacy toxenv setting.
@@ -321,7 +321,7 @@ class TestAfter:
             'toxenv = py36\n'
         )
         ini = py.iniconfig.IniConfig('', data=inistr)
-        assert after_config_matches(['py36'], ini)
+        assert after_config_matches(ini, ['py36'])
         out, err = capsys.readouterr()
         msg = 'The "toxenv" key of the [travis:after] section is deprecated'
         assert msg in err
@@ -339,7 +339,7 @@ class TestAfter:
             'toxenv = py36\n'
         )
         ini = py.iniconfig.IniConfig('', data=inistr)
-        assert not after_config_matches(['py35'], ini)
+        assert not after_config_matches(ini, ['py35'])
         out, err = capsys.readouterr()
         msg = 'The "toxenv" key of the [travis:after] section is deprecated'
         assert msg in err
@@ -354,7 +354,7 @@ class TestAfter:
             'envlist = py36\n'
         )
         ini = py.iniconfig.IniConfig('', data=inistr)
-        assert after_config_matches(['py36'], ini)
+        assert after_config_matches(ini, ['py36'])
 
     def test_after_config_matches_envlist_nomatch(self):
         """Test that it doesn't work."""
@@ -366,7 +366,7 @@ class TestAfter:
             'envlist = py36\n'
         )
         ini = py.iniconfig.IniConfig('', data=inistr)
-        assert not after_config_matches(['py35'], ini)
+        assert not after_config_matches(ini, ['py35'])
 
     def test_after_config_matches_env_match(self, monkeypatch):
         """Test that it works."""
@@ -383,7 +383,7 @@ class TestAfter:
             '    DJANGO: 1.11\n'
         )
         ini = py.iniconfig.IniConfig('', data=inistr)
-        assert after_config_matches(['py36'], ini)
+        assert after_config_matches(ini, ['py36'])
 
     def test_after_config_matches_env_nomatch(self, monkeypatch):
         """Test that it doesn't work."""
@@ -400,4 +400,4 @@ class TestAfter:
             '    DJANGO: 1.11\n'
         )
         ini = py.iniconfig.IniConfig('', data=inistr)
-        assert not after_config_matches(['py35'], ini)
+        assert not after_config_matches(ini, ['py35'])
